Move duplicate `MockSelectedSiteSettings` and `MockPOSSiteSettingService` mocks to be reusable.
