+++ WooCommerce/WooCommerceTests/ViewRelated/Settings/POS/POSTabEligibilityCheckerTests.swift
@@ -540,19 +540,6 @@ private extension POSTabEligibilityCheckerTests {
     }
 }
 
-private final class MockSelectedSiteSettings: SelectedSiteSettingsProtocol {
-    var mockSettingsStream: AnyPublisher<(siteID: Int64, settings: [SiteSetting], source: SettingsUpdateSource), Never>?
-    var siteSettings: [SiteSetting] = []
-
-    var settingsStream: AnyPublisher<(siteID: Int64, settings: [SiteSetting], source: SettingsUpdateSource), Never> {
-        return mockSettingsStream ?? Empty().eraseToAnyPublisher()
-    }
-
-    func refresh() {
-        // Mock implementation - no action needed.
-    }
-}
-
 private final class MockPOSSystemStatusService: POSSystemStatusServiceProtocol {
     var resultToReturn: Result<POSPluginAndFeatureInfo, Error> = .success(POSPluginAndFeatureInfo(wcPlugin: nil, featureValue: nil))
 
@@ -560,19 +547,6 @@ private final class MockPOSSystemStatusService: POSSystemStatusServiceProtocol {
         switch resultToReturn {
         case .success(let info):
             return info
-        case .failure(let error):
-            throw error
-        }
-    }
-}
-
-private final class MockPOSSiteSettingService: POSSiteSettingServiceProtocol {
-    var setFeatureResult: Result<Bool, Error> = .success(true)
-
-    func setFeature(siteID: Int64, feature: SiteSettingsFeature, enabled: Bool) async throws -> Bool {
-        switch setFeatureResult {
-        case .success(let result):
-            return result
         case .failure(let error):
             throw error
         }
